Use breakpad_getcontext on all Linux platforms missing getcontext
getcontext is also not available on musl libc, so generalize
breakpad_getcontext so it can be used as a fallback for non-Android
platforms as well.

On x86_64, ucontext_t uses an Android-specific offset for storage
of FP registers, since its sigset_t is 8 bytes instead of 128 bytes.
So, make the definition of MCONTEXT_FPREGS_MEM conditional on whether
we are building for Android.

On glibc and musl, signal.h and asm/sigcontext.h can't be included
together, so in breakpad_context_unittest.cc, only compare the libc
and kernel _fpstate when on Android.

Bug: google-breakpad:631